// src/Services/models/ApplicationTypeIEnumerableStandardResponse.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

import type { ApplicationType } from './ApplicationType';

export type ApplicationTypeIEnumerableStandardResponse = {
    href?: string | null;
    relations?: Array<string> | null;
    method?: string | null;
    routeName?: string | null;
    routeValues?: any;
    status?: boolean;
    message?: string | null;
    data?: Array<ApplicationType> | null;
    statusCode?: string | null;
    errors?: any;
};